requirements.txt
absl-py==2.3.1
accelerate==1.2.1
aiosignal==1.4.0
albucore==0.0.17
albumentations==1.4.18
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
av==12.3.0
banal==1.0.6
blessings==1.7
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
dataset==1.6.2
decord==0.6.0
diffusers==0.30.2
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.8.1
einops-exts==0.0.4
eval_type_backport==0.3.0
exceptiongroup==1.3.0
Farama-Notifications==0.0.4
fastapi==0.128.0
fastparquet==2024.11.0
filelock==3.20.0
flash-attn==2.7.1.post4
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
gr00t==1.1.0
greenlet==3.3.0
grpcio==1.76.0
gymnasium==1.0.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
imageio==2.34.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.3.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.15.0
kiwisolver==1.4.9
kornia==0.7.4
kornia_rs==0.1.10
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.45.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.62.1
numpy==1.26.4
numpydantic==1.6.7
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.6.1
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.18.0
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
peft==0.17.0
pettingzoo==1.25.0
pillow==12.0.0
pip==25.3
pipablepytorch3d==0.7.6
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
protobuf==4.25.1
psutil==7.1.3
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
ray==2.40.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.45.0
setproctitle==1.3.7
setuptools==80.9.0
shtab==1.8.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==1.4.54
starlette==0.50.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
threadpoolctl==3.6.0
tianshou==0.5.1
tifffile==2025.5.10
timm==1.0.14
tokenizers==0.21.4
tomli==2.3.0
torch==2.5.1
TorchCodec==0.1.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.51.3
triton==3.1.0
typeguard==4.4.2
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2025.2
umap-learn==0.5.11
urllib3==2.5.0
uvicorn==0.40.0
wandb==0.18.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.2
yacs==0.1.8
zipp==3.23.0
